+++ src/App.php
@@ -2,9 +2,9 @@
 
 namespace Kenjiefx\ScratchPHP;
 
+use Kenjiefx\ScratchPHP\App\Interfaces\ModuleInterface;
 use Kenjiefx\ScratchPHP\App\Modules\CLIModule;
 use Kenjiefx\ScratchPHP\App\Modules\PreviewModule;
-use Kenjiefx\ScratchPHP\App\Modules\ModuleInterface;
 
 class App
 {
